fix: refactor task construction to use builder pattern, allowing extensibility to build systems and simplified collector attachment.
