+++ src/main/java/org/eisopux/diagnostics/core/Collector.java
@@ -2,30 +2,19 @@ package org.eisopux.diagnostics.core;
 
 import java.util.List;
 
-import javax.tools.JavaCompiler;
-
 public interface Collector<T> {
 
     /**
      * Called before the javac compilation task is executed. Use this to register any TaskListeners
      * or initialize data structures.
      */
-    default void onBeforeCompile() {
-        // Default no-op
-    }
-
-    default void attachToTask(JavaCompiler.CompilationTask task) {
-        // Default no-op
-    }
+    default void onBeforeCompile(CompilationTaskBuilder builder) {}
 
     /**
-     * Called after the javac compilation task has finished.
-     *
-     * @param success {@code true} if the compilation succeeded; otherwise {@code false}.
+     * Called after compilation. The collector should finalize its data and contribute its section
+     * into the given report.
      */
-    default void onAfterCompile(boolean success) {
-        // Default no-op
-    }
+    default void onAfterCompile() {}
 
     List<T> getItems();
 }
