+++ src/main/java/org/eisopux/diagnostics/core/Reporter.java
@@ -5,10 +5,6 @@ import java.util.List;
 /** A pluggable interface for generating output from collected data. */
 public interface Reporter {
 
-    /**
-     * Produce a formatted report given:
-     *
-     * @param collectors The active collectors after compilation
-     */
+    /** Produce a formatted report given: */
     void generateReport(List<? extends Collector<?>> collectors);
 }
