+++ src/main/java/org/eisopux/diagnostics/reporter/ConsoleReporter.java
@@ -5,13 +5,22 @@ import org.eisopux.diagnostics.core.Reporter;
 
 import java.util.List;
 
-/** A very simple reporter that outputs raw collector data to the console. */
 public class ConsoleReporter implements Reporter {
 
     @Override
     public void generateReport(List<? extends Collector<?>> collectors) {
+        System.out.println("=== Compilation Report ===");
+
         for (Collector<?> collector : collectors) {
             System.out.println(collector.getItems());
         }
+        //        // Iterate over each section contributed by collectors.
+        //        reportData.getAllSections().forEach((sectionName, sectionData) -> {
+        //            System.out.println("Section: " + sectionName);
+        //            sectionData.forEach((key, value) -> {
+        //                System.out.println("  " + key + ": " + value);
+        //            });
+        //            System.out.println(); // Blank line between sections.
+        //        });
     }
 }
